Cleanup of test 'test_query_rules_fast_routing_algorithm-t' utilities
 - Moved new tap utility functions to 'utils.h'.
 - Removed previously borrowed utilities from another PR.
