+++ test/tap/tap/utils.h
@@ -11,6 +11,7 @@
 
 #include "curl/curl.h"
 #include <mysql.h>
+#include "sqlite3db.h"
 
 #include "command_line.h"
 
@@ -483,5 +484,38 @@ enum LINE_MATCH_T { POS, LINE, MATCHES };
  * @return All the lines found matching the regex.
  */
 std::vector<line_match_t> get_matching_lines(std::fstream& f_stream, const std::string& regex);
+
+/**
+ * @brief Opens a sqlite3 db file located in the supplied path with the provided flags.
+ * @param f_path Path to the 'db' file.
+ * @param db Output parameter to the sqlite3 object to be opened.
+ * @param flags The flags used to open the file.
+ * @return EXIT_SUCCESS in case of success, EXIT_FAILURE otherwise. Error cause is logged.
+ */
+int open_sqlite3_db(const std::string& f_path, sqlite3** db, int flags);
+
+using sq3_col_def_t = std::string;
+using sq3_row_t = std::vector<std::string>;
+using sq3_err_t = std::string;
+using sq3_res_t = std::tuple<std::vector<sq3_col_def_t>,std::vector<sq3_row_t>,int64_t,sq3_err_t>;
+
+enum SQ3_RES_T {
+	SQ3_COLUMNS_DEF,
+	SQ3_ROWS,
+	SQ3_AFFECTED_ROWS,
+	SQ3_ERR
+};
+
+/**
+ * @brief Executes the provided query in the supplied sqlite3 db object.
+ * @param db Already initialized 'sqlite3' handler.
+ * @param query The query to be executed.
+ * @return An 'sq3_result_t' object holding the result, depending on the type of query and result, different
+ *  fields will be populated, in case of success:
+ *   - For DQL stmts COLUMN_DEF and ROWS will hold the columns definitions and the rows from the resultset.
+ *   - For DML stmts the AFFECTED_ROWS will show the number of modified rows.
+ *  In case of failure, ERR field will be populated and others will remain empty.
+ */
+sq3_res_t sqlite3_execute_stmt(sqlite3* db, const std::string& query);
 
 #endif // #define UTILS_H
